Move RecommendedExtensions and KnowledgeBase into coupons directory.
