+++ plugins/woocommerce-admin/client/marketing/coupons/knowledge-base/index.js
