+++ plugins/woocommerce-admin/client/marketing/coupons/knowledge-base/placeholder.js
