+++ plugins/woocommerce-admin/client/marketing/coupons/knowledge-base/style.scss
