+++ plugins/woocommerce-admin/client/marketing/coupons/recommended-extensions/item.js
@@ -8,7 +8,7 @@ import { recordEvent } from '@woocommerce/tracks';
  * Internal dependencies
  */
 import './style.scss';
-import { ProductIcon } from '../../components/';
+import { ProductIcon } from '../../components';
 import { getInAppPurchaseUrl } from '../../../lib/in-app-purchase';
 
 const RecommendedExtensionsItem = ( {
