+++ plugins/woocommerce-admin/client/marketing/coupons/recommended-extensions/placeholder.js
