+++ plugins/woocommerce-admin/client/marketing/coupons/recommended-extensions/style.scss
