+++ plugins/woocommerce-admin/client/marketing/coupons/recommended-extensions/test/index.js
